requirements.txt
fastapi
uvicorn
langchain
openai
faiss-cpu
pydantic
python-multipart
langchain-community
pydantic-settings
langchain-openai
chromadb
pinecone-client 
langchain-pinecone
